feat(athena): moving athena types to shared
Ref: #2170
