+++ packages/shared/src/interface/external/athenahealth/cx-mapping.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+
+export const athenaSecondaryMappingsSchema = z.object({
+  departmentIds: z.string().array(),
+});
+export type AthenaSecondaryMappings = z.infer<typeof athenaSecondaryMappingsSchema>;
