+++ packages/shared/src/interface/external/athenahealth/index.ts
@@ -1,2 +1,3 @@
 export * from "./patient";
+export * from "./cx-mapping";
 export * from "./jwt-token";
