+++ packages/shared/src/interface/external/athenahealth/jwt-token.ts
@@ -1,6 +1,4 @@
-import { z } from "zod";
-
-export const athenaSecondaryMappingsSchema = z.object({
-  departmentIds: z.string().array(),
-});
-export type AthenaSecondaryMappings = z.infer<typeof athenaSecondaryMappingsSchema>;
+export type AthenaJwtTokenData = {
+  ah_practice: string;
+  ah_department: string;
+};
